# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tickets/customer": {
      "post": {
        "operationId": "TicketsController_createCustomerTicket",
        "summary": "Create a customer ticket linked to a booking",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer ticket created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/SupportTicket"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/tickets/provider": {
      "post": {
        "operationId": "TicketsController_createProviderTicket",
        "summary": "Create a provider support ticket",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider ticket created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/SupportTicket"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/tickets/my-tickets": {
      "get": {
        "operationId": "TicketsController_myTickets",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/tickets/my-provider-tickets": {
      "get": {
        "operationId": "TicketsController_myProviderTickets",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/tickets/{id}": {
      "get": {
        "operationId": "TicketsController_getTicket",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/support/tickets": {
      "get": {
        "operationId": "TicketsController_supportTickets",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/support/tickets/{id}": {
      "get": {
        "operationId": "TicketsController_supportTicket",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/support/tickets/{id}/status": {
      "patch": {
        "operationId": "TicketsController_supportStatus",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/support/tickets/{id}/remarks": {
      "patch": {
        "operationId": "TicketsController_supportRemarks",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketRemarksDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/support/tickets/{id}/resolve": {
      "patch": {
        "operationId": "TicketsController_supportResolve",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketRemarksDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/support/tickets/{id}/escalate": {
      "patch": {
        "operationId": "TicketsController_supportEscalate",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketRemarksDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/support/tickets/{id}/priority": {
      "patch": {
        "operationId": "TicketsController_supportPriority",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketPriorityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/admin/escalated-tickets": {
      "get": {
        "operationId": "TicketsController_adminEscalatedTickets",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/admin/tickets/{id}": {
      "get": {
        "operationId": "TicketsController_adminTicket",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/admin/tickets/{id}/admin-remarks": {
      "patch": {
        "operationId": "TicketsController_adminRemarks",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketRemarksDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/admin/tickets/{id}/final-decision": {
      "patch": {
        "operationId": "TicketsController_finalDecision",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketFinalDecisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/admin/tickets/{id}/resolve": {
      "patch": {
        "operationId": "TicketsController_resolve",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketFinalDecisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/admin/tickets/{id}/reject": {
      "patch": {
        "operationId": "TicketsController_reject",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use customer/user, provider, support, admin, or superuser role.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketFinalDecisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tickets"
        ]
      }
    }
  },
  "info": {
    "title": "ServeEase API",
    "description": "In-memory NestJS backend with CRUD, DTO validation, RBAC, and Swagger.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "Role-based access header. Use \"admin\" for full access or \"user\" for read-only GET access."
      }
    },
    "schemas": {
...
      "TicketStatusHistory": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Pending"
          },
          "note": {
            "type": "string",
            "example": "Ticket created."
          },
          "updatedBy": {
            "type": "string",
            "example": "Customer"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-05-20T10:30:00.000Z"
          }
        },
        "required": [
          "status",
          "note",
          "updatedBy",
          "updatedAt"
        ]
      },
      "SupportTicket": {
        "type": "object",
        "properties": {
          "ticketId": {
            "type": "string",
            "example": "TKT-20260520-1001"
          },
          "raisedByType": {
            "type": "string",
            "example": "customer",
            "enum": [
              "customer",
              "provider"
            ]
          },
          "raisedById": {
            "type": "string",
            "example": "CUS001"
          },
          "raisedByName": {
            "type": "string",
            "example": "Raghava Kumar"
          },
          "raisedByEmail": {
            "type": "string",
            "example": "user@serveease.com"
          },
          "raisedByPhone": {
            "type": "string",
            "example": "9876543210"
          },
          "relatedBookingId": {
            "type": "string",
            "example": "BOOK-2026-1045"
          },
          "relatedPaymentId": {
            "type": "string",
            "example": "TXN-2026-4582"
          },
          "providerId": {
            "type": "string",
            "example": "cleanpro-services"
          },
          "providerName": {
            "type": "string",
            "example": "CleanPro Services"
          },
          "customerId": {
            "type": "string",
            "example": "CUS001"
          },
          "customerName": {
            "type": "string",
            "example": "Raghava Kumar"
          },
          "ticketType": {
            "type": "string",
            "example": "Late Arrival"
          },
          "subject": {
            "type": "string",
            "example": "Provider arrived late"
          },
          "description": {
            "type": "string",
            "example": "Provider arrived 90 minutes late and service was rushed."
          },
          "attachmentUrl": {
            "type": "string",
            "example": "late-proof.jpg"
          },
          "priority": {
            "type": "string",
            "example": "High"
          },
          "status": {
            "type": "string",
            "example": "Pending"
          },
          "supportRemarks": {
            "type": "string",
            "example": "Verified booking details."
          },
          "adminRemarks": {
            "type": "string",
            "example": "Refund approved."
          },
          "escalationReason": {
            "type": "string",
            "example": "Refund approval required."
          },
          "finalDecision": {
            "type": "string",
            "example": "Refund Approved"
          },
          "assignedSupportId": {
            "type": "string",
            "example": "SUP001"
          },
          "assignedSupportName": {
            "type": "string",
            "example": "Priya Sharma"
          },
          "resolvedBy": {
            "type": "string",
            "example": "Support"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-05-20T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-05-20T10:30:00.000Z"
          },
          "resolvedAt": {
            "type": "string",
            "example": "2026-05-20T11:30:00.000Z"
          },
          "escalatedAt": {
            "type": "string",
            "example": "2026-05-20T11:00:00.000Z"
          },
          "statusHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketStatusHistory"
            }
          }
        },
        "required": [
          "ticketId",
          "raisedByType",
          "raisedById",
          "raisedByName",
          "raisedByEmail",
          "raisedByPhone",
          "ticketType",
          "subject",
          "description",
          "priority",
          "status",
          "createdAt",
          "updatedAt",
          "statusHistory"
        ]
      },
      "CreateCustomerTicketDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "example": "BOOK-2026-1045"
          },
          "ticketType": {
            "type": "string",
            "example": "Late Arrival",
            "enum": [
              "Poor Service Quality",
              "Provider Not Arrived",
              "Late Arrival",
              "Overcharging",
              "Misbehavior",
              "Payment Issue",
              "Service Not Completed",
              "Refund Request",
              "Fraud",
              "Safety Concern",
              "Other"
            ]
          },
          "subject": {
            "type": "string",
            "example": "Provider arrived late"
          },
          "description": {
            "type": "string",
            "example": "Provider arrived very late and did not complete the promised service."
          },
          "attachmentUrl": {
            "type": "string",
            "example": "delay-proof.jpg"
          },
          "customerId": {
            "type": "string",
            "example": "CUS001"
          },
          "customerName": {
            "type": "string",
            "example": "Raghava Kumar"
          }
        },
        "required": [
          "bookingId",
          "ticketType",
          "description"
        ]
      },
      "CreateProviderTicketDto": {
        "type": "object",
        "properties": {
          "ticketType": {
            "type": "string",
            "example": "Payment Not Received",
            "enum": [
              "Payment Not Received",
              "Verification Issue",
              "Booking Assignment Issue",
              "Technical Problem",
              "Customer Misbehavior",
              "Profile Update Request",
              "Availability Issue",
              "Account Suspension Appeal",
              "Service Category Change",
              "Other"
            ]
          },
          "subject": {
            "type": "string",
            "example": "Payment not received for completed booking"
          },
          "description": {
            "type": "string",
            "example": "Payout for BOOK-2026-1101 is still pending."
          },
          "relatedBookingId": {
            "type": "string",
            "example": "BOOK-2026-1101"
          },
          "relatedPaymentId": {
            "type": "string",
            "example": "TX-2026-7854"
          },
          "priority": {
            "type": "string",
            "example": "Medium",
            "enum": [
              "Low",
              "Medium",
              "High",
              "Critical"
            ]
          },
          "attachmentUrl": {
            "type": "string",
            "example": "payment-proof.jpg"
          },
          "providerId": {
            "type": "string",
            "example": "PRO001"
          },
          "providerName": {
            "type": "string",
            "example": "CleanPro Services"
          },
          "customerName": {
            "type": "string",
            "example": "Raghava Kumar"
          },
          "service": {
            "type": "string",
            "example": "Kitchen Cleaning"
          }
        },
        "required": [
          "ticketType",
          "subject",
          "description"
        ]
      },
      "TicketStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "In Progress",
            "enum": [
              "In Progress"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "TicketRemarksDto": {
        "type": "object",
        "properties": {
          "remarks": {
            "type": "string",
            "example": "Checked booking and contacted both parties."
          }
        },
        "required": [
          "remarks"
        ]
      },
      "TicketPriorityDto": {
        "type": "object",
        "properties": {
          "priority": {
            "type": "string",
            "example": "High",
            "enum": [
              "Low",
              "Medium",
              "High",
              "Critical"
            ]
          }
        },
        "required": [
          "priority"
        ]
      },
      "TicketFinalDecisionDto": {
        "type": "object",
        "properties": {
          "finalDecision": {
            "type": "string",
            "example": "Refund Approved",
            "enum": [
              "Refund Approved",
              "Refund Rejected",
              "Rework Scheduled",
              "Provider Warning Issued",
              "Provider Suspended",
              "Ticket Rejected",
              "No Action Required"
            ]
          },
          "adminRemarks": {
            "type": "string",
            "example": "Refund approved after support verified the service failure."
          }
        },
        "required": [
          "finalDecision",
          "adminRemarks"
        ]
      }
    }
  }
}